+++ src/commands/index.ts
@@ -160,6 +160,7 @@ import { cacheUpdate } from './moderation/cacheUpdate';
 import { navigraph } from './general/navigraph';
 import { botIssue } from './support/botIssue';
 import { slowMode } from './moderation/slowmode';
+import { econnreset } from './support/econnreset';
 import { pr } from './utils/pr';
 
 const commands: BaseCommandDefinition[] = [
@@ -323,6 +324,7 @@ const commands: BaseCommandDefinition[] = [
     navigraph,
     botIssue,
     slowMode,
+    econnreset,
     pr,
 ];