+++ .github/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Update <small>_ February 2023</small>
 
+- feat: ECONNRESET support command (07/02/2023)
 - fix: add link to docs to .trythis command (07/02/2023)
 - feat: PR command (07/02/2023)